Wrote reward function
Wrote the reward function, along with tests to make sure it works as expected.
The reward function operates on a very particular set of parameters, which may
not coincide with exactly what robosuite provides. As a result, the preprocess
function call needs to be used to convert how robosuite provides information
to how the reward function expects information.