Working reward function
I finished the reward function. I verified it works. There are still 2 potential
things I might add. The first will deny rewards if the camera is not held
horizontally (for the purpose of not producing upside-down videos).
The other is a flag that allows for the positive rewards to take on continuous
values based on how centered the block is on the screen. Either way,
this should now be complete enough for training.